requirements.txt
asgiref==3.4.0
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
base-url==0.0.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.3
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-base-url==2020.12.3
django-cors-headers==3.7.0
django-csv-export-view==1.1.0
django-import-export==2.5.0
django-model-utils==4.1.1
django-multiselectfield==0.1.12
django-rest-knox==4.1.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
heroku==0.1.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
kiwisolver==1.3.1
MarkupPy==1.14
matplotlib==3.3.4
numpy==1.21.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.1
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
roslibpy==1.2.1
rospy-message-converter==0.5.2
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
